+++ Assets/LDtkUnity/Editor/CustomEditor/SectionDrawers/LDtkSectionDrawer.cs
@@ -1,6 +1,4 @@
-﻿using System.Collections.Generic;
-using System.Linq;
-using UnityEditor;
+﻿using UnityEditor;
 using UnityEngine;
 
 namespace LDtkUnity.Editor