+++ Assets/LDtkUnity/Editor/ScriptedImporter/LDtkJsonImporter.cs
@@ -10,20 +10,25 @@ namespace LDtkUnity.Editor
 {
     public abstract class LDtkJsonImporter<T> : ScriptedImporter where T : ScriptableObject, ILDtkJsonFile
     {
-        protected T ReadAssetText(AssetImportContext ctx)
+        protected T ReadAssetText()
         {
-            string path = ctx.assetPath;
-            string json = File.ReadAllText(path);
+            string json = File.ReadAllText(assetPath);
 
             T file = ScriptableObject.CreateInstance<T>();
 
-            file.name = Path.GetFileNameWithoutExtension(ctx.assetPath);
+            file.name = Path.GetFileNameWithoutExtension(assetPath);
 
             //Debug.Log($"Reimporting {file.name}");
             
             file.SetJson(json);
 
             return file;
+        }
+
+        protected LdtkJson ReadJson()
+        {
+            string json = File.ReadAllText(assetPath);
+            return LdtkJson.FromJson(json);
         }
     }
 }