+++ Assets/LDtkUnity/CHANGELOG.md
@@ -1,11 +1,12 @@
 # 2.0.5
-###### June XX, 2021
+###### June 13, 2021
 - Added a warning dialog if an LDtk project/level tries to get moved
+- Added a safeguard for using Aseprite files, which are unsupported at the moment
+  - Added a new Aseprite sample scene
 - Adjusted the main section's Pixels Per Unit field
   - Value will never go below 1
   - Value will default to the `DefaultGridSize` in the LDtk json upon first import or reset
-- Fixed icon size for the enum namespace field if the namespace is invalid
-- Added a safeguard for using Aseprite files, which are unsupported at the moment
+- Fixed small icon size for the enum namespace field if the namespace is invalid
 
 # 2.0.4
 ###### June 6, 2021